Rename reCLAMM to AutoRange in UI and tooling
Rebrand the product to AutoRange across the React app, Firebase
function, and docs. Smart-contract references (ABI file, exported
const, on-chain identifiers) stay as reCLAMM since they're tied to
deployed bytecode. The /reclamm route now redirects to /autorange
for backward compatibility.
