+++ client/src/pool-types/autorange/AutoRangeMath.ts
