+++ client/src/pool-types/autorange/AutoRangePriceBar.tsx
@@ -5,9 +5,9 @@ import {
   calculateLowerMargin,
   calculateUpperMargin,
   computeCenteredness,
-} from './ReClammMath';
+} from './AutoRangeMath';
 
-interface ReClammPriceBarProps {
+interface AutoRangePriceBarProps {
   realTimeBalanceA: number;
   realTimeBalanceB: number;
   realTimeVirtualBalances: {
@@ -18,7 +18,7 @@ interface ReClammPriceBarProps {
   margin: number;
 }
 
-export const ReClammPriceBar: React.FC<ReClammPriceBarProps> = ({
+export const AutoRangePriceBar: React.FC<AutoRangePriceBarProps> = ({
   realTimeBalanceA,
   realTimeBalanceB,
   realTimeVirtualBalances,
@@ -206,7 +206,7 @@ export const ReClammPriceBar: React.FC<ReClammPriceBarProps> = ({
         style={{ marginBottom: 16, textAlign: 'center' }}
         ref={containerRef}
       >
-        ReClamm Price Range
+        AutoRange Price Range
       </Typography>
       <div style={{ display: 'flex', justifyContent: 'center' }}>
         <svg ref={svgRef} width={containerWidth} height='120' />
